Added workout service, added global error handler, workout tests and simple validation

+++ api/src/main/java/com/qualle/shapeup/api/ChartDto.java
@@ -15,5 +15,5 @@ public class ChartDto {
     private String description;
     private String measureType;
     private String imageLink;
-    private Map<String, Integer> data;
+    private Map<String, Float> data;
 }
